src/requirements.txt
Flask==2.0.1
python-dotenv==1.0.0
requests==2.31.0
python-telegram-bot==22.0
APScheduler==3.11.0
wikipedia-api==0.6.0
selenium==4.14.0
openai==0.28.1
google-genai==1.12.1
duckduckgo_search==7.5.1
Wikipedia-API==0.6.0
meteomatics==2.11.4
pymongo[srv]